+++ R/tg_categories.R
@@ -1,7 +1,7 @@
 #' Category list
 #' @description List of 'TGStat' channel categories
 #' @param lang Response language
-#' @references See also \href{https://api.tgstat.ru/docs/ru/database/categories.html}{TGstat API Documentation of metrod database/categories}
+#' @references See also \href{https://api.tgstat.ru/docs/ru/database/categories.html}{TGStat API Documentation of metrod database/categories}
 #' @return tibble with categories
 #' @export
 tg_categories <- function(
@@ -11,7 +11,8 @@ tg_categories <- function(
   data <- tg_make_request(
     method = 'database/categories',
     token  = tg_get_token(),
-    lang   = lang
+    lang   = lang,
+    check_quote = FALSE
   ) %>%
     tg_parse_response()