+++ R/tg_countries.R
@@ -4,7 +4,7 @@
 #'
 #' @return tibble with countries
 #' @export
-#' @references See also \href{https://api.tgstat.ru/docs/ru/database/countries.html}{TGstat API Documentation of metrod database/countries}
+#' @references See also \href{https://api.tgstat.ru/docs/ru/database/countries.html}{TGStat API Documentation of metrod database/countries}
 tg_countries <- function(
   lang = NULL
 ) {
@@ -12,7 +12,8 @@ tg_countries <- function(
   data <- tg_make_request(
     method = 'database/countries',
     token  = tg_get_token(),
-    lang   = lang
+    lang   = lang,
+    check_quote = FALSE
   ) %>%
     tg_parse_response()