+++ R/tg_languages.R
@@ -4,7 +4,7 @@
 #'
 #' @return tibble
 #' @export
-#' @references See also \href{https://api.tgstat.ru/docs/ru/database/languages.html}{TGstat API Documentation of metrod database/languages}
+#' @references See also \href{https://api.tgstat.ru/docs/ru/database/languages.html}{TGStat API Documentation of metrod database/languages}
 tg_languages <- function(
   lang = NULL
 ) {
@@ -12,7 +12,8 @@ tg_languages <- function(
   data <- tg_make_request(
     method = 'database/languages',
     token  = tg_get_token(),
-    lang   = lang
+    lang   = lang,
+    check_quote = FALSE
   ) %>%
     tg_parse_response()